moving component button to own folder
